+++ testdrivenapp/services/client/package.json
@@ -6,6 +6,7 @@
     "axios": "^0.18.1",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0"
   },
   "scripts": {
